add commit checker action
