+++ Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 name = "omnect-device-service"
 readme = "README.md"
 repository = "git@github.com:omnect/omnect-device-service.git"
-version = "0.17.1"
+version = "0.18.0"
 
 [dependencies]
 actix-server = "2.3"
